Tymy code

+++ src/Controller/CampusController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/campus')]
 final class CampusController extends AbstractController
@@ -17,6 +18,9 @@ final class CampusController extends AbstractController
     #[Route(name: 'app_campus_index', methods: ['GET'])]
     public function index(CampusRepository $campusRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         return $this->render('campus/index.html.twig', [
             'campuses' => $campusRepository->findAll(),
         ]);
@@ -25,6 +29,9 @@ final class CampusController extends AbstractController
     #[Route('/new', name: 'app_campus_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         $campus = new Campus();
         $form = $this->createForm(CampusType::class, $campus);
         $form->handleRequest($request);
@@ -45,6 +52,9 @@ final class CampusController extends AbstractController
     #[Route('/{id}', name: 'app_campus_show', methods: ['GET'])]
     public function show(Campus $campus): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         return $this->render('campus/show.html.twig', [
             'campus' => $campus,
         ]);
@@ -53,6 +63,9 @@ final class CampusController extends AbstractController
     #[Route('/{id}/edit', name: 'app_campus_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Campus $campus, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         $form = $this->createForm(CampusType::class, $campus);
         $form->handleRequest($request);
 
@@ -71,6 +84,9 @@ final class CampusController extends AbstractController
     #[Route('/{id}', name: 'app_campus_delete', methods: ['POST'])]
     public function delete(Request $request, Campus $campus, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         if ($this->isCsrfTokenValid('delete'.$campus->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($campus);
             $entityManager->flush();
