+++ src/Controller/CityController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/city')]
 final class CityController extends AbstractController
@@ -17,6 +18,9 @@ final class CityController extends AbstractController
     #[Route(name: 'app_city_index', methods: ['GET'])]
     public function index(CityRepository $cityRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         return $this->render('city/index.html.twig', [
             'cities' => $cityRepository->findAll(),
         ]);
@@ -25,6 +29,9 @@ final class CityController extends AbstractController
     #[Route('/new', name: 'app_city_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         $city = new City();
         $form = $this->createForm(CityType::class, $city);
         $form->handleRequest($request);
@@ -45,6 +52,9 @@ final class CityController extends AbstractController
     #[Route('/{id}', name: 'app_city_show', methods: ['GET'])]
     public function show(City $city): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         return $this->render('city/show.html.twig', [
             'city' => $city,
         ]);
@@ -53,6 +63,9 @@ final class CityController extends AbstractController
     #[Route('/{id}/edit', name: 'app_city_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, City $city, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         $form = $this->createForm(CityType::class, $city);
         $form->handleRequest($request);
 
@@ -71,6 +84,9 @@ final class CityController extends AbstractController
     #[Route('/{id}', name: 'app_city_delete', methods: ['POST'])]
     public function delete(Request $request, City $city, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé : vous devez être admin.');
+        }
         if ($this->isCsrfTokenValid('delete'.$city->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($city);
             $entityManager->flush();
